+++ test/unit/machine/code_coverage.sh
@@ -10,7 +10,6 @@ echo ""
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=machine test/unit/machine/linux.py
 coverage run -a --source=machine test/unit/machine/amachine.py
-coverage run -a --source=machine test/unit/machine/solaris.py
 
 echo ""
 echo "Producing code coverage report"